refactor: :truck: Moved utils module to modules folder
